feat: update renderFootprint to include ExtrudedPads in the model
