+++ lib/vanilla/index.ts
@@ -3,9 +3,10 @@ import { ExtrudedPads } from "../ExtrudedPads"
 import { h, Fragment, type VNode } from "./h"
 import { render, type RenderResult, type ColoredGeom } from "./render"
 export * from "./convertCSGToThreeGeom"
+export * from "./render"
 import type * as jscadModeling from "@jscad/modeling"
 
-export { h, Fragment }
+export { h, Fragment, Footprinter3d, ExtrudedPads }
 export type { VNode, RenderResult, ColoredGeom }
 
 export function getJscadModelForFootprint(
@@ -13,19 +14,6 @@ export function getJscadModelForFootprint(
   jscad: typeof jscadModeling,
 ): RenderResult {
   const vnode = h(Footprinter3d, { footprint })
-  return render(vnode, jscad)
-}
-
-export function getJscadModelForFootprintWithPads(
-  footprint: string,
-  jscad: typeof jscadModeling,
-): RenderResult {
-  const vnode = h(
-    Fragment,
-    null,
-    h(Footprinter3d, { footprint }),
-    h(ExtrudedPads, { footprint }),
-  )
   return render(vnode, jscad)
 }
